+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # v2.11.5
-## mm/dd/2018
+## 02/16/2018
 
 1. [](#new)
     * Added support for `form: process: - call: ['Class', 'method']` for custom form handling